requirements.txt
featurecloud
bios
torchvision
